# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Siam U Todo API
  version: 0.2.0
  description: Go + SQLite backend for authentication and user-owned Todo items.
servers:
  - url: http://localhost:4000
paths:
  /health:
    get:
      responses:
        '200': { description: Service is healthy }
  /api/auth/register:
    post:
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/RegisterInput' } }
      responses:
        '201': { description: Account created }
        '400': { description: Invalid input }
        '409': { description: Email already registered }
  /api/auth/login:
    post:
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/LoginInput' } }
      responses:
        '200': { description: Login successful }
        '401': { description: Invalid credentials }
  /api/auth/mock:
    post:
      description: Development-only mock OAuth provider.
      responses:
        '201': { description: Login successful }
  /api/auth/forgot-password:
    post:
      responses:
        '202': { description: Generic reset response }
  /api/me:
    get:
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Current user }
  /api/todos:
    get:
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: query, name: q, schema: { type: string } }
        - { in: query, name: status, schema: { type: string, enum: [completed, pending] } }
      responses:
        '200': { description: Owned Todo list }
    post:
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/TodoInput' } }
      responses:
        '201': { description: Todo created }
        '400': { description: Invalid Todo }
  /api/todos/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    get:
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Owned Todo }, '404': { description: Not found } }
    put:
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/TodoInput' } }
      responses: { '200': { description: Todo updated }, '404': { description: Not found } }
    delete:
      security: [{ bearerAuth: [] }]
      responses: { '204': { description: Todo deleted }, '404': { description: Not found } }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
    RegisterInput:
      type: object
      required: [email, name, password]
      properties:
        email: { type: string, format: email }
        name: { type: string }
        password: { type: string, minLength: 8 }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    TodoInput:
      type: object
      required: [title]
      properties:
        title: { type: string, maxLength: 240 }
        description: { type: string }
        completed: { type: boolean }
        priority: { type: integer, minimum: 0, maximum: 3 }